move MessageStatus to commons
